commited on 1/07/2026

+++ auth-service (1)/auth-service/src/main/java/com/example/auth_service/dto/UserResponse.java
@@ -6,22 +6,20 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class UserResponse {
-    private int id;
+    private Long id;
     private String username;
     private String email;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
-
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
     public String getUsername() {
         return username;
     }
-
     public void setUsername(String username) {
         this.username = username;
     }
@@ -29,7 +27,6 @@ public class UserResponse {
     public String getEmail() {
         return email;
     }
-
     public void setEmail(String email) {
         this.email = email;
     }
